[Rename] 컴포넌트 파스칼케이스명으로 변경
